UI improvements and bug fixing

+++ HCI_Manifestations/HCI_Manifestations/Dialogs/Manifestation/ShowManifestations.xaml.cs
@@ -99,7 +99,11 @@ namespace HCI_Manifestations.dialogs
 
         private void buttonDelete_Click(object sender, RoutedEventArgs e)
         {
-            Database.DeleteManifestation(SelectedManifestation);
+            if (SelectedManifestation != null)
+            {
+                Database.DeleteManifestation(SelectedManifestation);
+                SelectedManifestation = null;
+            }
         }
 
         private void manifestationsGrid_SelectionChanged(object sender, SelectionChangedEventArgs e)
